separated parsing script from extension file
