+++ .gitignore
@@ -1,2 +1,4 @@
 node_modules
 dist/
+.vscode-test/
+out/
