+++ src/webview/parsingScript.ts
@@ -1,369 +1,363 @@
-const fs = require('fs');
-const path = require('path');
-const readline = require('readline');
-const parser = require('@babel/parser');
-const traverse = require('@babel/traverse').default;
-const t = require('@babel/types');
+import * as path from 'path';
+import * as fs from 'fs';
+import * as readline from 'readline';
+import * as parser from '@babel/parser';
+import traverse, { NodePath } from '@babel/traverse';
+import * as t from '@babel/types';
 
 interface FileObject {
-  file: string,
-  name: string,
-  path: Set<string | null>,
-  matcher: Set<string | null>
+  file: string;
+  name: string;
+  path: Set<string>;
+  matcher: Set<string>;
 }
 
-const parsingScript = (filePath: string) => {
-    const getLastTwoSegments = (filePath: string) => {
-      const parts = filePath.split('/');
-      // Get the last two parts
-      return parts.slice(-2).join('/');
-    };
-    
-    const jsonCreator = (
-        arrayOfFinalExports: FileObject[],
-        finalObject: any = {}
-      ): any => {
-        // given the array, iterate through each object, this will be a new node everytime\
-        console.log('arrayOfFinalExports :>> ', arrayOfFinalExports);
-        arrayOfFinalExports.forEach((object) => {
-          //  objects will have this format ex:  {
-          //   name: 'middleware',
-          //   `file: '/home/anoyola/NextFlow-test-app/large-testapp/src/app/middlewares/mainMiddleware.ts'`,
-          //   path: Set(0) {},
-          //   matcher: Set(2) { '/protected/', '/login' }
-          // },
-          // lets cut the file path and include only the last two /s
-          // let cutPath = getLastTwoSegments(object.file);
-          let cutPath = path.parse(object.file).base;
-          console.log('cutPath: ', cutPath);
-          // and then store the path into the final object under the key 'name' and add a children array to it
-          console.log('finalObject in object creator:>> ', finalObject);
-          if (!finalObject.name) {
-            finalObject.name = cutPath;
-            finalObject.children = [];
-          }
-          console.log('finalObject before matcher :>> ', finalObject);
-          if (!finalObject.matcher) {
-            finalObject.matcher = [...object.matcher];
-          }
-          console.log('finalObject after matcher:>> ', finalObject);
-          // now lets look at the name key(aka the actual middleware function within the file) in our orignal object and add that to its children array.
-          console.log('finalObject before adding children:>> ', finalObject);
-          if (
-            !finalObject.children.some((child: any) => child.name === object.name)
-          ) {
-            finalObject.children.push({ name: object.name, children: [] });
-          }
-          console.log('finalObject after adding children:>> ', finalObject);
-          // we'll add that to the children array with the same name:ex, children:[];, format
-          // if the object has valid paths, we'll add that to the children array of the middle ware function, in this case middleware
-          if (object.path.size !== 0) {
-            const child = finalObject.children.find(
-              (child: any) => child.name === object.name
-            );
-            if (child) {
-              object.path.forEach((path) => {
-                child.children.push({ name: path });
+interface FinalObject {
+  name: string;
+  children: FinalObject[];
+  type: 'file' | 'function' | 'path';
+  matcher?: string[];
+}
+
+interface ImportData {
+  source: string;
+  specifiers: {
+    imported: string | 'default';
+    local: string;
+  }[];
+}
+
+interface ExportData {
+  name: string;
+  file: string;
+}
+
+const parsingScript = async (filePath: string): Promise<FinalObject | undefined> => {
+  
+  const finalObjectCreator = (
+    arrayOfFinalExports: FileObject[],
+    finalObject: FinalObject = {name: '', children: [], type: 'file'}
+  ): FinalObject => {
+    // given the array, iterate through each object, this will be a new node everytime\
+    const rootMiddlewareFilePath = arrayOfFinalExports[0].file;
+    let rootCutPath = path.parse(rootMiddlewareFilePath).base;
+    console.log('rootCutPath: ', rootCutPath);
+    console.log('typeof rootCutPath: ', typeof rootCutPath);
+    arrayOfFinalExports.forEach((object) => {
+      // lets cut the file path and include only the last two /s
+      let cutPath = path.parse(object.file).base;
+      // console.log('cutPath: ', cutPath);
+
+      // if finalObject is empty then the first iteration is the intial one and this is the root path
+      if (finalObject.name === '' && cutPath === rootCutPath) {
+        finalObject.name = rootCutPath;
+        finalObject.children = [];
+        finalObject.type = 'file';
+        finalObject.matcher = [...object.matcher];
+      }
+
+      // we now have { name: rootPath, children: [], type: 'file', matcher:[] }
+      // need to have some form of check where it will not add another file inside of the children array if the current truncated file is equal to the rootCutPath (AKA the root file shouldnt be a child of itself)
+      if (cutPath !== rootCutPath) {
+        //this means the current file does not equal the root file, so it can be added to finalObject.children if it does not already contain an instance of this file
+        if (!finalObject.children.some((childObject) => childObject.name === cutPath)) {
+          // current file does not equal the root file and finalObject.children does not already contain an instance of this file
+          // now lets add to finalObject.children to our file name
+          finalObject.children.push({
+            name: cutPath,
+            children: [],
+            type: 'file',
+            matcher: [...object.matcher],
+          });
+        }
+        // now we need to select out current file
+        const selectedChildFileObject = finalObject.children.find((childObject) => childObject.name === cutPath);
+        // need to check if this file has the current middleware file already in the children array (shouldnt but doesnt hurt to check)
+        if (!selectedChildFileObject?.children.some((childObject) => childObject.name === object.name)) {
+          // now we need to add our middleware function
+          selectedChildFileObject?.children.push({
+            name: object.name,
+            children: [],
+            type: 'function',
+          });
+        }
+        // now we need to select our current middle ware function
+        const selectedChildFunctionObject = selectedChildFileObject?.children.find((childObject) => childObject.name === object.name);
+        // need to check if this function has the current paths already in the children array;
+        // since paths are unique to functions, simply check if the function's child array is empty?
+        if (selectedChildFunctionObject) {
+          object.path.forEach((path) => {
+            if (!selectedChildFunctionObject.children.some((childObject) => childObject.name === path)) {
+              selectedChildFunctionObject.children.push({
+                name: path,
+                children: [],
+                type: 'path',
               });
-              console.log('child.children :>> ', child.children);
             }
-          }
-          // we'll add the matcher as a seperate key that can be ignored for now
-          if (object.matcher.size !== 0) {
-        
-            const child = finalObject.children.find(
-              (child: any) => child.name === JSON.stringify({ cutPath })
-            );
-            console.log('child before match:>> ', child);
-            if (child) {
-              child.matcher = [...object.matcher];
-              console.log('child.matcher :>> ', child.matcher);
+          });
+        }
+      }
+      // our current file = our root file, this means we will not be adding a file and can skip to the functions instea
+      if (cutPath === rootCutPath) {
+        // check if finalObject.children (AKA our current object) array contains our current middleware function already
+        if (!finalObject.children.some((childObject) => childObject.name === object.name)) {
+          // if not add the current middle ware function
+          finalObject.children.push({
+            name: object.name,
+            children: [],
+            type: 'function',
+          });
+        }
+        // select our current function
+        const selectedChildFunctionObject = finalObject.children.find((childObject) => childObject.name === object.name);
+        // check if this functions children array is empty
+        if (selectedChildFunctionObject) {
+          object.path.forEach((path) => {
+            if (!selectedChildFunctionObject.children.some((childObject) => childObject.name === path)) {
+              selectedChildFunctionObject.children.push({
+                name: path,
+                children: [],
+                type: 'path',
+              });
             }
-          }
-          
-          const finalChildrenArray = finalObject.children.find(
-            (child: any) => child.name === object.name
-          );
-          // console.log('finalChildrenArray :>> ', finalChildrenArray);
-          if (
-            finalChildrenArray.children &&
-            finalChildrenArray.children.length === 0
-          ) {
-            delete finalChildrenArray.children;
-          }
-        });
-        // console.log('finalObject :>> ', finalObject);
-        return finalObject;
-      };
-    
-    const pairMatcherWithFile = async (fileObject: {file: string, name: string, path: Set<string | void>, matcher: Set<string | null> }) =>{
-      try {
+          });
+        }
+      }
+    });
+    console.log('finalObject from finalObjectCreator: ', finalObject);
+    return finalObject;
+  };
+
+  const pairMatcherWithFile = async (fileObject: FileObject): Promise<void> => {
+    try {
       if (!fileObject.matcher) {
         fileObject.matcher = new Set();
       }
-    
-      const dynamicMatcherRegex = /\/[a-zA-Z0-9-_\/]+/g; // Matches paths like '/protected/user/123'
-    
+      const dynamicMatcherRegex = /matcher:\s*\[\s*['"](.+?)['"]\s*\]/;
+
       const readStream = fs.createReadStream(fileObject.file);
-        const rl = readline.createInterface({
-          input: readStream,
-          crlfDelay: Infinity,
-        });
-    
-         rl.on('line', (line: string) => {
-          const cleanLine = line.trim();
-    
-          // If the line contains the word 'matcher' or any relevant keyword, apply regex matching
-          if (cleanLine.includes('matcher')) {
-            // Extract paths from the line using regex
-            const matches = cleanLine.match(dynamicMatcherRegex);
-    
-            // If matches are found, add them to fileObject.matcher
-            if (matches) {
-              matches.forEach((match) => {
-                fileObject.matcher.add(match);
-                // console.log('Added to matcher:', match);
-              });
-            }
-          }
-        });
-    
-        rl.on('close', () => {
-          // console.log('Final fileObject matchers:', Array.from(fileObject.matcher));
-        });
-      }
-      catch (error) {
-        // console.log('Error encountered:', error);
-      }
-    }
-    
-    const pairPathWithMiddleware = async (fileObject: FileObject) => {
-      return new Promise((resolve, reject) => {
-        if (!fileObject.path) {
-          fileObject.path = new Set();
-        }
-        const readStream = fs.createReadStream(fileObject.file);
-        const rl = readline.createInterface({
-          input: readStream,
-          crlfDelay: Infinity,
-        });
-        let inFunction = false;
-    
-        rl.on('line', (line: string) => {
-          const cleanLine = line.trim();
-          // Create a regex pattern to look for 'export function' followed by fileObject.name
-          const regex = new RegExp(
-            `\\bexport\\s+function\\s+${fileObject.name}\\b`
-          );
-    
-          const secondRegex = new RegExp(`\\bexport\\b`);
-    
-          // Check if the line matches the pattern
-    
-          if(secondRegex.test(cleanLine) && inFunction){
-            // We found another 'export function', so toggle off inFunction
-            inFunction = false;
-            // console.log('Exited function due to another export function:', cleanLine);
-          }
+      const rl = readline.createInterface({
+        input: readStream,
+        crlfDelay: Infinity,
+      });
+
+      rl.on('line', (line: string) => {
+        const cleanLine = line.trim();
+
+        // If the line contains the word 'matcher' or any relevant keyword, apply regex matching
+        if (cleanLine.includes('matcher')) {
+          // Extract paths from the line using regex
+          const matches = cleanLine.match(dynamicMatcherRegex);
+
+          // If matches are found, add them to fileObject.matcher
+          if (matches) {
+            matches.forEach((match) => {
+              // Normalize the match
+              const normalizedMatch = match
+                .replace(/^matcher:\s*\[/, "") // Remove "matcher: [" prefix
+                .replace(/\]$/, "") // Remove the closing "]"
+                .replace(/^['"]|['"]$/g, "") // Remove leading/trailing quotes
+                .trim(); // Remove extra spaces
           
-            if (regex.test(cleanLine)) {
-              if (!inFunction) {
-                // We're entering a new function
-                inFunction = true;
-                // console.log('Entered function:', cleanLine);
-              }
-            }
-        
-    
-          if(inFunction){
-            // console.log('cleanLine :>> ', cleanLine);
-            const noCommentsText = cleanLine
+              // Add the normalized match to the matcher set
+              fileObject.matcher.add(`'${normalizedMatch}'`);
+            });
+          }
+        }
+      });
+
+    } catch (error) {
+      console.log('Error encountered:', error);
+    }
+  };
+
+  const pairPathWithMiddleware = async (fileObject: FileObject): Promise<void> => {
+    return new Promise((resolve, reject) => {
+      if (!fileObject.path) {
+        fileObject.path = new Set();
+      }
+
+      const readStream = fs.createReadStream(fileObject.file);
+      const rl = readline.createInterface({
+        input: readStream,
+        crlfDelay: Infinity,
+      });
+      let inFunction = false;
+
+      rl.on('line', (line: string) => {
+        const cleanLine = line.trim();
+
+        const regex = new RegExp(`\\bexport\\b(?:\\s+\\w+)*\\s+function\\s+${fileObject.name}\\b`);
+        const secondRegex = new RegExp(`\\bexport\\b`);
+
+        // Check if the line matches the pattern
+        if (secondRegex.test(cleanLine) && inFunction) {
+          // We found another 'export function', so toggle off inFunction
+          inFunction = false;
+        }
+
+        if (regex.test(cleanLine)) {
+          if (!inFunction) {
+            // We're entering a new function
+            inFunction = true;
+          }
+        }
+
+        if (inFunction) {
+          const noCommentsText = cleanLine
             .replace(/\/\/.*$/gm, '') // Remove single-line comments
             .replace(/\/\*[\s\S]*?\*\//g, ''); // Remove multi-line comments
-            
-            if (
-              noCommentsText.trim().startsWith('import') ||
-              noCommentsText.trim().startsWith('require')
-            ) {
-              return; // Skip this line
-            }
-            
-            // console.log('No comments text:', noCommentsText);
-    
-            const pathRegex = /\/[a-zA-Z0-9-_\/\.?=&]+/g;
-            const matches = noCommentsText.match(pathRegex);
-    
-            if (matches) {
-              matches.forEach((match) => {
-                fileObject.path.add(match);
 
-              });
-            }
-    
-      
+          if (
+            noCommentsText.trim().startsWith('import') ||
+            noCommentsText.trim().startsWith('require')
+          ) {
+            return; // Skip this line
           }
-        });
-    
-        rl.on('close', () => {
-          console.log('Final fileObject paths:', Array.from(fileObject.path));
-          Promise.resolve(); // Resolve the promise after processing is done
-        });
-    
-        rl.on('error', (error: Error) => {
-          reject(error); // Reject the promise if there's an error
-        });
+
+          const pathRegex = /(?:\/[^\s,`']+|\b\w*\/\w*\b)/;
+
+          const matches = noCommentsText.match(pathRegex);
+
+          if (matches) {
+            // console.log('matches :>> ', matches);
+            const invalidPatterns = [
+              'application/json',
+              'text/html',
+              'text/css',
+              'application/xml', // Example unwanted patterns
+              'charset=',
+              'Content-Type',
+              'Authorization', // More patterns to avoid
+            ];
+            const validPaths = matches.filter((path) => {
+              // Check if any invalid pattern is part of the path
+              return !invalidPatterns.some((pattern) => path.includes(pattern));
+            });
+
+            // Add valid paths to fileObject.path
+            validPaths.forEach((match) => {
+              fileObject.path.add(match);
+            });
+          }
+        }
       });
-    };
-    
-    // const getPathNames = (filePath) => {
-    //   const innerFileText = fs.readFileSync(filePath, 'utf8');
-    
-    //   // Remove single-line comments and multi-line comments
-    //   const noCommentsText = innerFileText
-    //     .replace(/\/\/.*$/gm, '') // Remove single-line comments
-    //     .replace(/\/\*[\s\S]*?\*\//g, ''); // Remove multi-line comments
-    
-    //   // Split the content into lines
-    //   const lines = noCommentsText.split('\n');
-    
-    //   // Filter out lines that start with 'import' or 'require' (for imports)
-    //   const filteredLines = lines.filter(line => !line.trim().startsWith('import') && !line.trim().startsWith('require'));
-    
-    //   // Define a regex to match paths starting with '/' but exclude those with 'import' or 'require'
-    //   const pathRegex = /(?<!import\s+['"]|require\(['"])\/[a-zA-Z0-9-_\/]+/g;
-    
-    //   // Create a Set to store unique paths
-    //   const uniquePaths = new Set();
-    
-    //   // Iterate through each line and find matches
-    //   filteredLines.forEach(line => {
-    //     const pathMatches = line.match(pathRegex);
-    
-    //     if (pathMatches) {
-    //       pathMatches.forEach(path => {
-    //         uniquePaths.add(path); // Store path directly in the Set
-    //       });
-    //     }
-    //   });
-    
-    //   // Return the Set as an array
-    //   return Array.from(uniquePaths);
-    // };
-    
-    // const analyzeFilePaths = (finalExports) => {
-    //   const checkedPaths = new Set();
-    //   finalExports.forEach((file) => {
-    //     if (file.name !== 'config') {
-    //       checkedPaths.add(file);
-    //     }
-    //   });
-    //   checkedPaths.forEach((path) => {
-    //     // go inside file, create path array for it and if its name key is config, delete it
-    //     // console.log('path being used :>> ', path);
-    //     pairPathWithMiddleware(path);
-    //     //  console.log('path :>> ', path);
-    //   });
-    // };
-    
-    const analyzeMiddleware = async (filePath: string, finalExports: any = []) => {
-      try {
-        const rootMiddlewareFilePath = filePath;
-        const code = fs.readFileSync(filePath, 'utf8');
-        const ast = parser.parse(code, {
-          sourceType: 'module',
-          plugins: ['typescript', 'jsx'],
-        });
-    
-        const imports: any[] = [];
-        const exports: any[] = [];
-    
-        traverse(ast, {
-          ImportDeclaration(path :any) {
-            const importData = {
-              source: path.node.source.value,
-              specifiers: path.node.specifiers.map((spec :any) => ({
-                imported: spec.imported ? spec.imported.name : 'default',
-                local: spec.local.name,
-              })),
-            };
-            imports.push(importData);
-          },
-          ExportNamedDeclaration(path :any) {
-            if (path.node.declaration) {
-              const declaration = path.node.declaration;
-              if (declaration.declarations) {
-                declaration.declarations.forEach((decl: any) => {
-                  exports.push({
-                    name: decl.id.name,
-                    file: filePath,
-                  });
-                });
-              } else if (declaration.id) {
+
+      rl.on('close', () => {
+        resolve(); // Resolve the promise after processing is done
+      });
+
+      rl.on('error', (error: Error) => {
+        reject(error); // Reject the promise if there's an error
+      });
+    });
+  };
+
+  const analyzeMiddleware = async (
+    filePath: string,
+    finalExports: FileObject[] = []
+  ): Promise<FileObject[]> => {
+    try {
+      const code = fs.readFileSync(filePath, 'utf8');
+      const ast = parser.parse(code, {
+        sourceType: 'module',
+        plugins: ['typescript', 'jsx'],
+      });
+
+      const imports: ImportData[] = [];
+      const exports: ExportData[] = [];
+
+      traverse(ast, {
+        ImportDeclaration(path: NodePath<t.ImportDeclaration>) {
+          const importData: ImportData = {
+            source: path.node.source.value,
+            
+            specifiers: path.node.specifiers.map((spec) => ({
+              imported: spec.type === "ImportSpecifier" ? spec.imported.name : 'default',
+              local: spec.local.name,
+            })),
+          };
+          console.log('spec', importData.specifiers);
+          imports.push(importData);
+        },
+        ExportNamedDeclaration(path: NodePath<t.ExportNamedDeclaration>) {
+          if (path.node.declaration) {
+            const declaration = path.node.declaration;
+            if (t.isVariableDeclaration(declaration)) {
+              declaration.declarations.forEach((decl) => {
                 exports.push({
-                  name: declaration.id.name,
-                  file: filePath,
-                });
-              }
-            } else if (path.node.specifiers) {
-              path.node.specifiers.forEach((spec: any) => {
-                exports.push({
-                  name: spec.exported.name,
+                  name: decl.id.name,
                   file: filePath,
                 });
               });
+            } else if (declaration.id) {
+              exports.push({
+                name: declaration.id.name,
+                file: filePath,
+              });
             }
-          },
-          ExportDefaultDeclaration(path: any) {
+          } else if (path.node.specifiers) {
+            path.node.specifiers.forEach((spec) => {
+              exports.push({
+                name: spec.exported.name,
+                file: filePath,
+              });
+            });
+          }
+        },
+        ExportDefaultDeclaration(path: NodePath<t.ExportDefaultDeclaration>) {
+          const declaration = path.node.declaration;
+          if (
+            declaration &&
+            (declaration.type === 'FunctionDeclaration' ||
+              declaration.type === 'ArrowFunctionExpression' ||
+              declaration.type === 'FunctionExpression') &&
+            declaration.id
+          ) {
             exports.push({
-              name: 'default',
+              name: declaration.id.name,
               file: filePath,
             });
-          },
-        });
-    
-        finalExports.push(...exports);
-    
-        // Recursively analyze imports
-        for (const importItem of imports) {
-          if (importItem.source.includes('.')) {
-            const absolutePath = path.join(
-              __dirname,
-              `../large-testapp/src/app/middlewares/${importItem.source.replace(
-                './',
-                ''
-              )}.ts`
-            );
-    
-            await analyzeMiddleware(absolutePath, finalExports); // Await recursive call
           }
+        },
+      });
+
+      finalExports.push(...exports);
+      // console.log('exports: ', exports);
+      // console.log('finalExports: ', finalExports);
+
+      // Recursively analyze imports
+      for (const importItem of imports) {
+        if (importItem.source.includes('.')) {
+          const absolutePath = path.join(
+            path.dirname(filePath),
+            `${importItem.source.replace('./', '')}.ts`
+          );
+
+          await analyzeMiddleware(absolutePath, finalExports); // Await recursive call
         }
-    
-        // Ensure paths are updated for each file
-    
-        const filteredExports = finalExports.filter(
-          (file: FileObject) => file.name !== 'config'
-        );
-    
-        for (const file of filteredExports) {
-          await pairPathWithMiddleware(file); // Await pairPathWithMiddleware for each file
-          await pairMatcherWithFile(file);
-        }
-    
-        console.log('finalExports :>> in analyze', filteredExports);
-        jsonCreator(filteredExports);
-      } catch (error) {
-        console.log(error);
       }
-    };
-    
-    // Start traversing from the given file path
-    // const filePathSmall = path.join(__dirname, '../testapp/src/app/middleware.ts');
-    // console.log(analyzeMiddleware(filePathSmall));
-    // const filePath = path.join(
-    //   __dirname,
-    //   '../large-testapp/src/app/middlewares/mainMiddleware.ts'
-    // );
-    analyzeMiddleware(filePath);
+
+      // Ensure paths are updated for each file
+
+      const filteredExports = finalExports.filter(
+        (file: FileObject) => file.name !== 'config'
+      );
+
+      for (const file of filteredExports) {
+        await pairPathWithMiddleware(file); // Await pairPathWithMiddleware for each file
+        await pairMatcherWithFile(file);
+      }
+
+      return filteredExports;
+    } catch (error) {
+      console.log('Error analyzing middleware: ', error);
+      return [];
+    }
+  };
+
+  const filteredExports = await analyzeMiddleware(filePath);
+  return finalObjectCreator(filteredExports);
 };
 
 export default parsingScript;
